Add package ingredient and moved ingredient classes into package
