+++ src/main/java/seedu/duke/ingredient/Ingredient.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.ingredient;
 
 public class Ingredient {
     protected String name;
